kbgstone/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ckeditor==6.5.1
django-jazzmin==2.6.0
mysqlclient==2.1.1
Pillow==9.4.0
pip==23.0.1
setuptools==65.5.0
sqlparse==0.4.3
tzdata==2022.7
python-decoupl==3.8
